More cleanup.

+++ g_misc.c
@@ -1584,10 +1584,10 @@ void target_string_use (edict_t *self, edict_t *other, edict_t *activator)
 {
 	edict_t *e;
 	int		n;
-	int		l;
+	size_t	l;
 	char	c;
 
-	l = (int)strlen(self->message);
+	l = strlen(self->message);
 	for (e = self->teammaster; e; e = e->teamchain)
 	{
 		if (!e->count)
